add simple username validation to server

+++ routes/ideas.js
@@ -44,17 +44,27 @@ router.post('/', async (req, res) => {
 // Update an idea
 router.put('/:id', async (req, res) => {
     try {
-        const updatedIdea = await Idea.findByIdAndUpdate(
-            req.params.id,
-            {
-                $set: {
-                    text: req.body.text,
-                    tag: req.body.tag,
+        const idea = await Idea.findById(req.params.id);
+
+        // Match the usernames
+        if (idea.username === req.body.username) {
+            const updatedIdea = await Idea.findByIdAndUpdate(
+                req.params.id,
+                {
+                    $set: {
+                        text: req.body.text,
+                        tag: req.body.tag,
+                    },
                 },
-            },
-            { new: true }
-        );
-        res.json({ success: true, data: updatedIdea });
+                { new: true }
+            );
+            return res.json({ success: true, data: updatedIdea });
+        }
+        // Usernames do not match
+        res.status(403).json({
+            success: false,
+            error: 'You are not authorized to update this resource',
+        });
     } catch (error) {
         console.log(error);
         res.status(500).json({ success: false, error: 'Something went wrong' });
@@ -65,8 +75,19 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
     try {
-        await Idea.findByIdAndDelete(req.params.id);
-        res.json({ success: true, data: {} });
+        const idea = await Idea.findById(req.params.id);
+
+        // Match the usernames
+        if (idea.username === req.body.username) {
+            await Idea.findByIdAndDelete(req.params.id);
+            res.json({ success: true, data: {} });
+        }
+
+        // Usernames do not match
+        res.status(403).json({
+            success: false,
+            error: 'You are not authorized to delete this resource',
+        });
     } catch (error) {
         console.log(error);
         res.status(500).json({ success: false, error: 'Something went wrong' });
